A few conversions from float to int and defining constants as float (rather than double) to reduce warnings with Intel comiler.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@2130 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/layers/Gratings.cpp
@@ -51,7 +51,7 @@ bool Gratings::updateImage(float time, float dt)
    for (int iy = 0; iy < ny; iy++) {
       for (int ix = 0; ix < nx; ix++) {
          float x = (float) ix;
-         data[ix*sx + iy*sy] = sin(kx * x + phi);
+         data[ix*sx + iy*sy] = sinf(kx * x + phi);
       }
    }
